// telegram-bot/gmx-sdk/gmx/domain/synthetics/trade/types.ts
export enum TriggerThresholdType {
  Above = ">",
  Below = "<",
}

export type SwapAmounts = {
  amountIn: BigNumber;
  usdIn: BigNumber;
  amountOut: BigNumber;
  usdOut: BigNumber;
  priceIn: BigNumber;
  priceOut: BigNumber;
  swapPathStats: SwapPathStats | undefined;
  minOutputAmount: BigNumber;
};

export type IncreasePositionAmounts = {
  initialCollateralAmount: BigNumber;
  initialCollateralUsd: BigNumber;

  collateralDeltaAmount: BigNumber;
  collateralDeltaUsd: BigNumber;

  swapPathStats: SwapPathStats | undefined;

  indexTokenAmount: BigNumber;

  sizeDeltaUsd: BigNumber;
  sizeDeltaInTokens: BigNumber;

  estimatedLeverage?: BigNumber;

  indexPrice: BigNumber;
  initialCollateralPrice: BigNumber;
  collateralPrice: BigNumber;
  triggerPrice?: BigNumber;
  triggerThresholdType?: TriggerThresholdType;
  acceptablePrice: BigNumber;
  acceptablePriceDeltaBps: BigNumber;

  positionFeeUsd: BigNumber;
  feeDiscountUsd: BigNumber;
  borrowingFeeUsd: BigNumber;
  fundingFeeUsd: BigNumber;
  positionPriceImpactDeltaUsd: BigNumber;
};

export type DecreasePositionAmounts = {
  isFullClose: boolean;
  sizeDeltaUsd: BigNumber;
  sizeDeltaInTokens: BigNumber;
  collateralDeltaUsd: BigNumber;
  collateralDeltaAmount: BigNumber;

  indexPrice: BigNumber;
  collateralPrice: BigNumber;
  triggerPrice?: BigNumber;
  acceptablePrice: BigNumber;
  acceptablePriceDeltaBps: BigNumber;

  estimatedPnl: BigNumber;
  estimatedPnlPercentage: BigNumber;
  realizedPnl: BigNumber;

  positionFeeUsd: BigNumber;
  feeDiscountUsd: BigNumber;
  borrowingFeeUsd: BigNumber;
  fundingFeeUsd: BigNumber;
  swapProfitFeeUsd: BigNumber;
  positionPriceImpactDeltaUsd: BigNumber;
  priceImpactDiffUsd: BigNumber;
  payedRemainingCollateralAmount: BigNumber;

  payedOutputUsd: BigNumber;
  payedRemainingCollateralUsd: BigNumber;

  receiveTokenAmount: BigNumber;
  receiveUsd: BigNumber;

  triggerOrderType?: OrderType.LimitDecrease | OrderType.StopLossDecrease;
  triggerThresholdType?: TriggerThresholdType;
  decreaseSwapType: DecreasePositionSwapType;
};
// ...
export type SwapStats = {
  marketAddress: string;
  tokenInAddress: string;
  tokenOutAddress: string;
  isWrap: boolean;
  isUnwrap: boolean;
  isOutLiquidity?: boolean;
  swapFeeAmount: BigNumber;
  swapFeeUsd: BigNumber;
  priceImpactDeltaUsd: BigNumber;
  amountIn: BigNumber;
  amountInAfterFees: BigNumber;
  usdIn: BigNumber;
  amountOut: BigNumber;
  usdOut: BigNumber;
};

export type SwapPathStats = {
  swapPath: string[];
  swapSteps: SwapStats[];
  targetMarketAddress?: string;
  totalSwapPriceImpactDeltaUsd: BigNumber;
  totalSwapFeeUsd: BigNumber;
  totalFeesDeltaUsd: BigNumber;
  tokenInAddress: string;
  tokenOutAddress: string;
  usdOut: BigNumber;
  amountOut: BigNumber;
};